refactor auth strategy
